+++ tests/integration/test_check_in.py
@@ -17,7 +17,7 @@ from lib.utils import BASE_URL
 def handler(mocker: MockerFixture) -> None:
     mock_scheduler = mocker.patch("lib.checkin_scheduler.CheckInScheduler")
     flight_info = {
-        "arrivalAirport": {"name": "test_inbound"},
+        "arrivalAirport": {"name": "test_inbound", "country": None},
         "departureAirport": {"code": "LAX", "name": "test_outbound"},
         "departureDate": "2021-12-06",
         "departureTime": "14:40",